moving_peak added
Still some problems here:
Need same() in class continuous
In original version constraint_calue() in problem is not pure virtual, but now it is. So i can't instantiate moving_peak, and I can't add a function unused in moving_peak neither. So maybe it should not be a pure virtual function.

Changed some names.